bug fixed in images Loading in detailFragment

+++ Remote/src/main/java/ir/omidtaheri/remote/entity/response/MovieImagesResponseRemoteEntity.kt
@@ -11,7 +11,7 @@ data class Backdrop(
     val file_path: String,
     val height: Int,
     val iso_639_1: String?,
-    val vote_average: Int,
+    val vote_average: Double,
     val vote_count: Int,
     val width: Int
 )
@@ -21,7 +21,7 @@ data class Poster(
     val file_path: String,
     val height: Int,
     val iso_639_1: String?,
-    val vote_average: Int,
+    val vote_average: Double,
     val vote_count: Int,
     val width: Int
 )
